Config to allow subset of processors and exporters
This adds two new job properties that let's an operator restrict the
kinds of OTel processors and exporters that can be used:
`allow_list.processors` and `allow_list.exporters`. These properties
are empty by default ([]), which keeps the existing behavior of
allowing all processors and exporters that are included in our
distribution. If not empty, then only the included components are
allowed.

Signed-off-by: Ivan Protsiuk <ivan.protsiuk@broadcom.com>
Signed-off-by: Matthew Kocher <matthew.kocher@broadcom.com>
Signed-off-by: Glenn Oppegard <glenn.oppegard@broadcom.com>