Add default Prometheus exporter configuration
This change introduces the ability to add an optional default Prometheus exporter to
the OpenTelemetry Collector via BOSH properties.

A new prom_exporter_config property allows defining a Prometheus exporter, which will
be added to the collector's configuration. This simplifies adding exporters without
needing to modify the base configuration directly. TLS for the exporter can be configured
using prom_exporter_tls properties.

The configuration logic ensures that the new exporter does not conflict with existing
ones on the same port and handles unique naming for exporters.